refactor: consolidate test utils. refactor: re-export testing-library/react functions with render override. refactor: use screen instead of destructured exports from render.
