+++ src/components/App.test.js
@@ -1,19 +1,18 @@
 import { server } from 'mocks/server';
 import { rest } from 'msw';
-import { renderWithProviders } from 'utils/test-utils';
 import App from './App';
-import { fireEvent } from '@testing-library/react';
+import { fireEvent, screen, render, givenServerRespondsWithSearchRequest } from 'test-utils';
 import selectEvent from 'react-select-event';
 
 it('renders without crashing', async () => {
-    renderWithProviders(<App />);
+    render(<App />);
 });
 
 it('renders with a default search result', async () => {
 
-    const {findByText} = renderWithProviders(<App />);
+    render(<App />);
 
-    expect(await findByText("example entry")).toBeInTheDocument();
+    expect(await screen.findByText("example entry")).toBeInTheDocument();
 
 });
 
@@ -27,10 +26,10 @@ it("renders an error banner if the logbook service cannot be reached", async () 
     );
 
     // When rendered
-    const {findByText} = renderWithProviders(<App />);
+    render(<App />);
 
     // Then an error message is present
-    expect(await findByText(/Search Error/i)).toBeInTheDocument();
+    expect(await screen.findByText(/Search Error/i)).toBeInTheDocument();
 
 });
 
@@ -47,11 +46,11 @@ it("displays sign-in and disabled log entry when logged out", async () => {
 
 
     // When rendered
-    const {findByText} = renderWithProviders(<App />);
+    render(<App />);
 
     // Then the user is logged out and cannot create log entries
-    expect(await findByText(/Sign In/i)).toBeInTheDocument();
-    expect(await findByText(/New Log Entry/i)).toBeDisabled();
+    expect(await screen.findByText(/Sign In/i)).toBeInTheDocument();
+    expect(await screen.findByText(/New Log Entry/i)).toBeDisabled();
 
 });
 
@@ -70,28 +69,28 @@ it("displays username and allows creating log entries when signed in", async () 
     );
 
     // When rendered
-    const {findByText} = renderWithProviders(<App />);
+    render(<App />);
 
     // Then the user is logged in and can create log entries
-    expect(await findByText(/garfieldHatesMondays/i)).toBeInTheDocument();
-    expect(await findByText(/New Log Entry/i)).toBeEnabled();
+    expect(await screen.findByText(/garfieldHatesMondays/i)).toBeInTheDocument();
+    expect(await screen.findByText(/New Log Entry/i)).toBeEnabled();
 
 });
 
 it("allows user to manually enter search terms", async () => {
 
     // Given app is rendered with default search results
-    const {findByText, findByDisplayValue} = renderWithProviders(<App />);
-    expect(await findByText("example entry")).toBeInTheDocument();
+    render(<App />);
+    expect(await screen.findByText("example entry")).toBeInTheDocument();
 
     // When user hits enter
     fireEvent.keyDown(
-        await findByDisplayValue(/start=12/),
+        await screen.findByDisplayValue(/start=12/),
         {key: 'Enter', code: 'Enter'}
     );
 
     // Then those search results are still there / unchanged
-    expect(await findByText("example entry")).toBeInTheDocument();
+    expect(await screen.findByText("example entry")).toBeInTheDocument();
 
     // And given the server will respond with new search results
     givenServerRespondsWithSearchRequest({
@@ -103,24 +102,24 @@ it("allows user to manually enter search terms", async () => {
     // enters new search terms,
     // and presses enter
     fireEvent.input(
-        await findByDisplayValue(/start=/),
+        await screen.findByDisplayValue(/start=/),
         {target: { value: 'start=987654321' }}
     )
     fireEvent.keyDown(
-        await findByDisplayValue(/start=987654321/),
+        await screen.findByDisplayValue(/start=987654321/),
         {key: 'Enter', code: 'Enter'}
     );
 
     // then the results are updated
-    expect(await findByText("hmmm title")).toBeInTheDocument();
+    expect(await screen.findByText("hmmm title")).toBeInTheDocument();
 
 });
 
 it("allows user to search with search filter bar", async () => {
 
     // Given app is rendered with default search results
-    const {findByText, findByLabelText} = renderWithProviders(<App />);
-    expect(await findByText("example entry")).toBeInTheDocument();
+    render(<App />);
+    expect(await screen.findByText("example entry")).toBeInTheDocument();
 
     // And given the server will respond with updated search results
     givenServerRespondsWithSearchRequest({
@@ -130,26 +129,26 @@ it("allows user to search with search filter bar", async () => {
 
     // When user opens the filter bar, updates the query, and closes it
     fireEvent.click(
-        await findByLabelText(/Show Search Filters/i)
+        await screen.findByLabelText(/Show Search Filters/i)
     )
     fireEvent.input(
-        await findByLabelText(/Title/i),
+        await screen.findByLabelText(/Title/i),
         {target: {value: 'some value'}}
     )
     fireEvent.click(
-        await findByLabelText(/Show Search Filters/i)
+        await screen.findByLabelText(/Show Search Filters/i)
     )
 
     // then the results are updated
-    expect(await findByText("hmmm title")).toBeInTheDocument();
+    expect(await screen.findByText("hmmm title")).toBeInTheDocument();
 
 });
 
 it("updates search results instantly from the search filter bar for tags", async () => {
 
     // Given app is rendered with default search results
-    const {findByText, findByLabelText} = renderWithProviders(<App />);
-    expect(await findByText("example entry")).toBeInTheDocument();
+    render(<App />);
+    expect(await screen.findByText("example entry")).toBeInTheDocument();
 
     // Given the server responds with updated search results
     givenServerRespondsWithSearchRequest({
@@ -159,21 +158,21 @@ it("updates search results instantly from the search filter bar for tags", async
 
     // When user opens the filter bar, updates the query, and closes it
     fireEvent.click(
-        await findByLabelText(/Show Search Filters/i)
+        await screen.findByLabelText(/Show Search Filters/i)
     )
     // unfortunately, atm this is react-select dependent
-    await selectEvent.select(await findByLabelText(/Tags/i), ['foo']);
+    await selectEvent.select(await screen.findByLabelText(/Tags/i), ['foo']);
 
     // then the results are updated
-    expect(await findByText("hmmm title")).toBeInTheDocument();
+    expect(await screen.findByText("hmmm title")).toBeInTheDocument();
 
 });
 
 it("updates search results instantly from the search filter bar for logbooks", async () => {
 
     // Given app is rendered with default search results
-    const {findByText, findByLabelText} = renderWithProviders(<App />);
-    expect(await findByText("example entry")).toBeInTheDocument();
+    render(<App />);
+    expect(await screen.findByText("example entry")).toBeInTheDocument();
 
     // When user opens the filter bar, and updates the query without closing it
     // (and we expect different results from the server)
@@ -183,54 +182,12 @@ it("updates search results instantly from the search filter bar for logbooks", a
     });
     
     fireEvent.click(
-        await findByLabelText(/Show Search Filters/i)
+        await screen.findByLabelText(/Show Search Filters/i)
     )
     // unfortunately, atm this is react-select dependent
-    await selectEvent.select(await findByLabelText(/Logbooks/i), ['controls']);
+    await selectEvent.select(await screen.findByLabelText(/Logbooks/i), ['controls']);
 
     // then the results are updated
-    expect(await findByText("hmmm title")).toBeInTheDocument();
+    expect(await screen.findByText("hmmm title")).toBeInTheDocument();
 
 });
-
-// Utility test function that will setup the server to respond to the request
-// with a log entry having the desired `title` is the `requestPredicate` is true
-// otherwise will respond with empty search results
-const givenServerRespondsWithSearchRequest = ({title, requestPredicate}) => {
-    server.use(
-        rest.get('*/logs/search', (req, res, ctx) => {
-            if(requestPredicate(req)) {
-                return res(
-                    ctx.json({
-                        hitCount: 1,
-                        logs: [
-                            {
-                                "id": 45,
-                                "owner": "jones",
-                                "source": title + " description",
-                                "description": title + " description",
-                                "title": title,
-                                "level": "Normal",
-                                "state": "Active",
-                                "createdDate": 1656599929021,
-                                "modifyDate": null,
-                                "events": null,
-                                "logbooks": [],
-                                "tags": [],
-                                "properties": [],
-                                "attachments": []
-                            }
-                        ]
-                    })
-                );
-            } else {
-                return res(
-                    ctx.json({
-                        hitCount: 0,
-                        logs: []
-                    })
-                );
-            }
-        })
-    )
-}
